Finished problem 5 of algorithms
Finished the fifth problem in the category algorithms. The DJ alg strikes again. It was a really interesting problem since we only had one circuit to work with, but 4 oracles, and a max of 8 qubits. Using the auxiliary qubits in creative and effective ways is definitely a skill to improve