Finished the last problem in qchem
Finishing the last problem in the category qchem: mind_the_gap. I was really excited to tackle this one because it was about implementing a VQE to find the ground state energy, and first excited energy level, for a system H-H (without spin). It was nice because I know this is something that's used in the QC industry quite a bit, and it was easy enough to implement.

In the spirit of QHACK, I stayed inside the #QHACK lines when editing the code. However, if I were doing this myself, I would implement some functions differently to save space since there were quite a few lines that were repeated in the two VQE functions.