Finished the second problem in games
Finally finished the second problem in games. It didn't take me weeks, but I did look at it a couple times. My final main issue that stumped me for half an hour or so was the cost function. I had it slightly wrong for the optimization alg I was using.