Finished problem 3 of games
Finished the problem labeled Elitzur Vaidman in the category games. Although my answer vs the given solution is concerning.

I find that my answer = (1 - answer file). For the input file, the answer is 0.5, so I was able to find the correct result. However, for the second problem, the given answer is 0.9. I am convinced this isn't correct, and the answer should be 0.1.

From what I understand, if you sample state 0 before the final measurement, there is a bomb and it explodes. If you measure state 0 in the final measurement, then detector D beeps, and if you measure state 1 in the final, detector C.

I am certain I have figured out how to measure those correctly, and my result turns out to be ~0.1 or 10% probability.

For my answer to match up with the answer file, I had to apply an X-gate to to the final step. This doesn't make sense to me and so I may have to revisit this problem in the future