Finished the fourth problem in games
I have finished the fourth problem in games called find_the_car.

After testing every case, it can be verified that my algorithm gets the answer correct 100% the time.

However, it does not flip the third qubit for the correct solution. Based on how I set up my post processing, the answer can always be found to be correct, but the states from the sample do not necessarily reflect the 'proper' way of finding the answer.