Finished last problem in Games
I quickly finished the last problem in the category games: game_switches.

Due to it's similarity, I started with solving it with the Bernstein Vazirani alg, and seeing where that would take me. It turns out that it is the correct way of doing it, and I didn't need to edit the problem further.