Finished problem 3 in qml
Finished problem 3 in qml, ising_classifier.

This one took me awhile because it seems that if you use the cost function inputs in a for loop, then python doesn't think they are being used to set the outputs. Had to change line 101 to what it is now, and that allowed my weights to update.